Inline upgrade credits source refinement
