+++ client/my-sites/plans-features-main/components/plan-notice-upgrade-credit.tsx
@@ -1,15 +1,66 @@
+import {
+	isDomainMapping,
+	isDomainRegistration,
+	isDomainTransfer,
+	isWpComPlan,
+} from '@automattic/calypso-products';
 import { formatCurrency } from '@automattic/number-formatters';
 import QuerySitePlans from 'calypso/components/data/query-site-plans';
 import QuerySitePurchases from 'calypso/components/data/query-site-purchases';
 import Notice from 'calypso/components/notice';
-import { getUpgradeCreditsEffectiveSource } from 'calypso/my-sites/plans-features-main/components/get-upgrade-credits-effective-source';
 import UpgradeCreditsNoticeText from 'calypso/my-sites/plans-features-main/components/upgrade-credits-notice-text';
 import { useUpgradeCreditsNoticeData } from 'calypso/my-sites/plans-features-main/hooks/use-upgrade-credits-notice';
 import { useSelector } from 'calypso/state';
 import { getCurrentUserCurrencyCode } from 'calypso/state/currency-code/selectors';
 import { getSitePurchases } from 'calypso/state/purchases/selectors/get-site-purchases';
+import type {
+	UpgradeCreditsNoticeData,
+	UpgradeCreditsNoticeSource,
+} from '../hooks/use-upgrade-credits-notice';
 import type { PlanSlug } from '@automattic/calypso-products';
 import type { PlansIntent } from '@automattic/plans-grid-next';
+import type { Purchase } from 'calypso/lib/purchases/types';
+
+function hasOtherUpgradesPurchase( sitePurchases: Purchase[] | undefined ): boolean {
+	return (
+		sitePurchases?.some( ( purchase ) => {
+			const productSlug = purchase?.productSlug;
+			if ( ! productSlug ) {
+				return false;
+			}
+
+			// "Other upgrades" means non-domain and non-plan purchases (e.g. themes add-on, storage, etc).
+			if ( isWpComPlan( productSlug ) ) {
+				return false;
+			}
+			if (
+				isDomainRegistration( purchase ) ||
+				isDomainTransfer( purchase ) ||
+				isDomainMapping( purchase )
+			) {
+				return false;
+			}
+
+			return true;
+		} ) ?? false
+	);
+}
+
+/**
+ * The upgrade credits hook can't always infer "domain + other upgrades" (some add-on related
+ * credits only appear as a generic non-coupon discount). When we can see purchases, refine
+ * the source so copy matches what contributed to the credit.
+ */
+function getUpgradeCreditsEffectiveSource(
+	upgradeCreditsNoticeData: UpgradeCreditsNoticeData | null,
+	sitePurchases: Purchase[] | undefined
+): UpgradeCreditsNoticeSource | undefined {
+	const source = upgradeCreditsNoticeData?.source;
+	if ( source === 'domain' && hasOtherUpgradesPurchase( sitePurchases ) ) {
+		return 'domain-and-other-upgrades';
+	}
+	return source;
+}
 
 type Props = {
 	className?: string;
